+++ packages/angular-meteor-data/.eslintignore
@@ -1,0 +1,5 @@
+dist
+lib
+tests
+webpack
+package.js
