+++ packages/angular-meteor-data/.eslintrc
@@ -1,0 +1,24 @@
+{
+  "extends": "airbnb/base",
+  "env": {
+    "browser": true
+  },
+  "globals": {
+    "angular": false,
+    "_": false,
+    "Meteor": false,
+    "Package": false,
+    "Tracker": false,
+    "jsondiffpatch": false
+  },
+  "rules": {
+    "quotes": [2, "single"],
+    "comma-dangle": 0,
+    "func-names": 0,
+    "no-param-reassign": 1,
+    "arrow-body-style": 0,
+    "space-before-function-paren": 0,
+    "no-use-before-define": [2, "nofunc"],
+    "brace-style": [1, "1tbs"]
+  }
+}
