--- packages/angular-meteor-data/modules/angular-meteor-settings.js
@@ -1,4 +1,0 @@
-angular.module('angular-meteor.settings', [])
-  .constant('$angularMeteorSettings', {
-    suppressWarnings: false
-  });
