+++ packages/angular-meteor-data/src/modules/angular-meteor-camera.js
