+++ packages/angular-meteor-data/src/modules/angular-meteor-collection.js
