+++ packages/angular-meteor-data/src/modules/angular-meteor-object.js
