+++ packages/angular-meteor-data/src/modules/angular-meteor-session.js
