+++ packages/angular-meteor-data/src/modules/angular-meteor-stopper.js
