+++ packages/angular-meteor-data/src/modules/angular-meteor-subscribe.js
