+++ packages/angular-meteor-data/src/modules/angular-meteor-user.js
