+++ packages/angular-meteor-data/src/modules/angular-meteor-utils.js
