+++ packages/angular-meteor-data/webpack/dist.js
@@ -1,0 +1,5 @@
+module.exports = require('./common')({
+  output: {
+    filename: '[name].js'
+  }
+});
