+++ packages/angular-meteor-data/webpack/prod.js
@@ -1,0 +1,8 @@
+module.exports = require('./common')({
+  output: {
+    // add .min suffix
+    filename: '[name].min.js'
+  },
+  // add source map
+  devtool: 'source-map'
+});
