+++ run_tests.sh
@@ -1,2 +1,2 @@
 cd packages/angular-meteor-data
-VELOCITY_TEST_PACKAGES=1 meteor test-packages --driver-package velocity:html-reporter@0.9.0 ./
+npm run test:watch
