+++ run_tests_ci.sh
@@ -1,2 +1,3 @@
 cd packages/angular-meteor-data
-VELOCITY_TEST_PACKAGES=1 meteor test-packages --velocity ./
+npm install
+npm test
